+++ src/model.ts
@@ -7,8 +7,3 @@ export type PremiseConfiguration = {
   cucumber?: CucumberConfiguration;
   version: 1;
 };
-
-export type ChangedRequirement = {
-  affects: string[];
-  id: string;
-};